python_bot/requirements.txt
alpaca-trade-api==3.2.0
google-generativeai==0.8.3
requests==2.32.3
pandas==2.2.2
chromadb==0.5.21
python-dotenv==1.0.1
